+++ packages/agent/src/agent/http/index.ts
@@ -53,8 +53,8 @@ export enum RequestStatusResponseStatus {
   Done = 'done',
 }
 
-// Default delta for ingress expiry is 5 minutes.
-const DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS = 5 * 60 * 1000;
+// Default delta for ingress expiry is 3 minutes.
+const DEFAULT_INGRESS_EXPIRY_DELTA_IN_MSECS = 3 * 60 * 1000;
 
 // Root public key for the IC, encoded as hex
 export const IC_ROOT_KEY =